pyproject.toml
[tool.poetry]
name = "pubsubgame"
version = "0.1.0"
description = ""
authors = ["AresAutomata <ares@automata.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nats-py = "^2.9.0"
pydantic = "^2.10.3"
pygame = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
